+++ python/rlgym_learn_algos/ppo/ppo_learner.py
@@ -253,6 +253,8 @@ class PPOLearner(
             ObsType,
             ActionType,
             RewardType,
+            ObsSpaceType,
+            ActionSpaceType,
             TrajectoryProcessorData,
         ],
     ):